+++ invokeai/frontend/web/src/features/nodes/components/fields/LoRAModelInputFieldComponent.tsx
@@ -1,7 +1,7 @@
 import { Flex, Text } from '@chakra-ui/react';
 import { SelectItem } from '@mantine/core';
 import { useAppDispatch } from 'app/store/storeHooks';
-import IAIMantineSelect from 'common/components/IAIMantineSelect';
+import IAIMantineSearchableSelect from 'common/components/IAIMantineSearchableSelect';
 import IAIMantineSelectItemWithTooltip from 'common/components/IAIMantineSelectItemWithTooltip';
 import { fieldValueChanged } from 'features/nodes/store/nodesSlice';
 import {
@@ -89,7 +89,7 @@ const LoRAModelInputFieldComponent = (
   }
 
   return (
-    <IAIMantineSelect
+    <IAIMantineSearchableSelect
       value={selectedLoRAModel?.id ?? null}
       label={
         selectedLoRAModel?.base_model &&
