+++ invokeai/frontend/web/src/features/parameters/components/Parameters/MainModel/ParamMainModelSelect.tsx
@@ -2,7 +2,7 @@ import { memo, useCallback, useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
 
 import { useAppDispatch, useAppSelector } from 'app/store/storeHooks';
-import IAIMantineSelect from 'common/components/IAIMantineSelect';
+import IAIMantineSearchableSelect from 'common/components/IAIMantineSearchableSelect';
 
 import { SelectItem } from '@mantine/core';
 import { createSelector } from '@reduxjs/toolkit';
@@ -77,14 +77,14 @@ const ParamMainModelSelect = () => {
   );
 
   return isLoading ? (
-    <IAIMantineSelect
+    <IAIMantineSearchableSelect
       label={t('modelManager.model')}
       placeholder="Loading..."
       disabled={true}
       data={[]}
     />
   ) : (
-    <IAIMantineSelect
+    <IAIMantineSearchableSelect
       tooltip={selectedModel?.description}
       label={t('modelManager.model')}
       value={selectedModel?.id}
