+++ invokeai/frontend/web/src/features/ui/components/tabs/ModelManager/subpanels/ModelManagerPanel/ModelListItem.tsx
@@ -1,7 +1,8 @@
 import { DeleteIcon } from '@chakra-ui/icons';
-import { Button, Flex, Text, Tooltip } from '@chakra-ui/react';
+import { Flex, Text, Tooltip } from '@chakra-ui/react';
 import { useAppSelector } from 'app/store/storeHooks';
 import IAIAlertDialog from 'common/components/IAIAlertDialog';
+import IAIButton from 'common/components/IAIButton';
 import IAIIconButton from 'common/components/IAIIconButton';
 import { selectIsBusy } from 'features/system/store/systemSelectors';
 import { useCallback } from 'react';
@@ -36,7 +37,8 @@ export default function ModelListItem(props: ModelListItemProps) {
   return (
     <Flex sx={{ gap: 2, alignItems: 'center', w: 'full' }}>
       <Flex
-        as={Button}
+        as={IAIButton}
+        isChecked={isSelected}
         sx={{
           justifyContent: 'start',
           p: 2,
